move ApiError to types
